Moved vehicle classes to separate package
