+++ src/trafficsimulator/vehicles/CautiousCar.java
@@ -3,21 +3,22 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package trafficsimulator.core;
+package trafficsimulator.vehicles;
 
+import trafficsimulator.core.Lane;
 import trafficsimulator.utils.Point;
 
 /**
  *
  * @author snorri
  */
-public class NormalBus extends Car {
+public class CautiousCar extends Car {
     
     private int topSpeed = 5;
     private int maxAcceleration = 1;
-    private String type = "Normal Bus";
+    private String type = "Cautious Car";
 
-    public NormalBus(Lane lane, Point position) {
+    public CautiousCar(Lane lane, Point position) {
         super(lane, position);
     }
