+++ src/trafficsimulator/vehicles/RecklessBus.java
@@ -3,21 +3,22 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package trafficsimulator.core;
+package trafficsimulator.vehicles;
 
+import trafficsimulator.core.Lane;
 import trafficsimulator.utils.Point;
 
 /**
  *
  * @author snorri
  */
-public class CautiousCar extends Car {
+public class RecklessBus extends Car {
     
-    private int topSpeed = 5;
-    private int maxAcceleration = 1;
-    private String type = "Cautious Car";
+    private int topSpeed = 10;
+    private int maxAcceleration = 2;
+    private String type = "Reckless Bus";
 
-    public CautiousCar(Lane lane, Point position) {
+    public RecklessBus(Lane lane, Point position) {
         super(lane, position);
     }
