+++ src/trafficsimulator/vehicles/RecklessCar.java
@@ -3,21 +3,22 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package trafficsimulator.core;
+package trafficsimulator.vehicles;
 
+import trafficsimulator.core.Lane;
 import trafficsimulator.utils.Point;
 
 /**
  *
  * @author snorri
  */
-public class NormalCar extends Car {
+public class RecklessCar extends Car {
     
-    private int topSpeed = 10;
-    private int maxAcceleration = 2;
-    private String type = "Normal Car";
+    private int topSpeed = 15;
+    private int maxAcceleration = 3;
+    private String type = "Reckless Car";
 
-    public NormalCar(Lane lane, Point position) {
+    public RecklessCar(Lane lane, Point position) {
         super(lane, position);
     }
